prevent premature posts saves from overwriting all panels
If a user manages to save a post in the few milliseconds before panel initialization
is able to disable form submission, the entire panel collection would be wiped out.
By setting a flag only when initilization is complete, we can verify that panels
were initialized and only then update the panels stored with the post.