settings fields - updates #61999
When adding a field to a panel type, you can use ->add_settings_field()
instead of ->add_field(). The former adds the field name to the settings
field list, which passes through to the blueprint.

The admin UI will use this data to separate fields into two tabs: Content and Settings