save empty values for all fields - updates #67595
When the user doesn't set a value for a field in the UI, the
react app doesn't event submit the field. This results in a lot
of missing keys in the database, and usually results int
notices in the templates.

This updates the saving process to make sure that all fields
have a sensible empty value set when the panel is saved with
no value submitted for that field