gracefully handle a missing post in a Post_List selection
Currently, if a post is manually selected in a Post_List field, and
is subsequently deleted, its slot in the Post_List permanently says
'Loading...', with no option to delete the post.

This addresses the issue in two ways:

1. Adds a record to the pre-cache for that post ID with the title
'Error: post not found', so that the user knows what happened.

2. Next time the panel is saved, removes the post from the list
automatically, freeing up the slot.